Interface code cleaning
