+++ client/internal/iface_windows.go
@@ -1,4 +1,4 @@
-package iface
+package internal
 
 import (
 	"net"
@@ -13,7 +13,7 @@ import (
 	"github.com/netbirdio/netbird/client/iface/wgproxy"
 )
 
-type IWGIface interface {
+type WGIface interface {
 	Create() error
 	CreateOnAndroid(routeRange []string, ip string, domains []string) error
 	IsUserspaceBind() bool
@@ -33,5 +33,6 @@ type IWGIface interface {
 	GetDevice() *device.FilteredDevice
 	GetWGDevice() *wgdevice.Device
 	GetStats(peerKey string) (configurer.WGStats, error)
+
 	GetInterfaceGUIDString() (string, error)
 }
